@@ -15,25 +15,26 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.3.1+1
+version: 0.3.2+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.13.4 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
+  plugin_platform_interface: ^2.0.1
   font_awesome_flutter: ^9.0.0
   shared_preferences: ^2.0.5
   http: ^0.13.1
   dart_rss: ^2.0.1
-  badges: ^1.2.0
+  badges: ^2.0.1
   intl: ^0.17.0
   google_fonts: ^2.0.0
   url_launcher: ^6.0.3
   modal_bottom_sheet: ^2.0.0
-  flutter_html: ^1.3.0
+  flutter_html: ^2.1.0
   jiffy: ^4.1.0
 
 
